Adapt placeholders to SMS
