+++ .devcontainer/.dev_config.yaml
@@ -1,4 +1,1 @@
-# Please only mention the non-default settings here:
-
-language: Greek
 service_instance_id: "1"
