--- .devcontainer/dev_launcher
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-# adapt to package name
-my-microservice
