+++ .pyproject_generation/pyproject_custom.toml
@@ -1,19 +1,15 @@
 [project]
-# please adapt to package name
-name = "my_microservice"
+name = "sms"
 version = "0.1.0"
-description = "My-Microservice - a short description"
+description = "State Management Service - Provides a REST API for basic infrastructure technology state management."
 dependencies = [
     "typer >= 0.12",
     "ghga-service-commons[api] >= 3.1",
-    "ghga-event-schemas >= 3.3",
-    "hexkit[akafka,s3,mongodb] >= 3.2",
+    "hexkit[mongodb] >= 3.2",
 ]
 
 [project.urls]
-# please adapt to package name
-Repository = "https://github.com/ghga-de/my-microservice"
+Repository = "https://github.com/ghga-de/state-management-service"
 
 [project.scripts]
-# please adapt to package name
-my-microservice = "my_microservice.__main__:run"
+sms = "sms.__main__:run"
