+++ .readme_generation/design.md
@@ -1,7 +1,16 @@
-<!-- Please provide an overview of the architecture and design of the code base.
-Mention anything that deviates from the standard triple hexagonal architecture and
-the corresponding structure. -->
+API calls to the State Management Service are authenticated through a configured API Key.
 
-This is a Python-based service following the Triple Hexagonal Architecture pattern.
-It uses protocol/provider pairs and dependency injection mechanisms provided by the
-[hexkit](https://github.com/ghga-de/hexkit) library.
+For each technology, REST API endpoints will be exposed, prefixed as follows:
+
+| **Technology Name** | **Prefix**  |
+|---------------------|-------------|
+| MongoDB             | `/docs/`    |
+| Apache Kafka        | `/events/`  |
+| S3                  | `/objects/` |
+| Vault               | `/secrets/` |
+
+Branch isolation is achieved in shared state technologies by the use of prefixes in,
+for example, database names.
+The `db_prefix` and `topic_prefix` config values are used so the prefixes must only be
+specified once. They are applied to database and topic names automatically throughout
+the SMS instance.
