+++ Dockerfile
@@ -44,5 +44,4 @@ WORKDIR /home/appuser
 USER appuser
 # set environment
 ENV PYTHONUNBUFFERED=1
-# Please adapt to package name:
-ENTRYPOINT ["my-microservice"]
+ENTRYPOINT ["sms"]
