--- example_data/README.md
@@ -1,4 +1,0 @@
-# Example Data
-This folder is may contain data (e.g. as json files) that can be used
-during development or testing, to set the application to an initial
-state, e.g. by populating the applications database with (realistic) data.
