+++ pyproject.toml
@@ -20,24 +20,23 @@ classifiers = [
     "Topic :: Software Development :: Libraries",
     "Intended Audience :: Developers",
 ]
-name = "my_microservice"
+name = "sms"
 version = "0.1.0"
-description = "My-Microservice - a short description"
+description = "State Management Service - Provides a REST API for basic infrastructure technology state management."
 dependencies = [
     "typer >= 0.12",
     "ghga-service-commons[api] >= 3.1",
-    "ghga-event-schemas >= 3.3",
-    "hexkit[akafka,s3,mongodb] >= 3.2",
+    "hexkit[mongodb] >= 3.2",
 ]
 
 [project.license]
 text = "Apache 2.0"
 
 [project.urls]
-Repository = "https://github.com/ghga-de/my-microservice"
+Repository = "https://github.com/ghga-de/state-management-service"
 
 [project.scripts]
-my-microservice = "my_microservice.__main__:run"
+sms = "sms.__main__:run"
 
 [tool.setuptools.packages.find]
 where = [
