+++ scripts/script_utils/fastapi_app_location.py
@@ -15,9 +15,3 @@
 #
 
 """Used to define the location of the main FastAPI app object."""
-
-# flake8: noqa
-# pylint: skip-file
-
-# Please adapt to package structure:
-from my_microservice.api.main import app
