+++ src/sms/__main__.py
@@ -15,18 +15,11 @@
 
 """Entrypoint of the package."""
 
-import asyncio
-
-from ghga_service_commons.api import run_server
-
-from .api.main import app  # noqa: F401 pylint: disable=unused-import
 from .config import CONFIG, Config
 
 
 def run(config: Config = CONFIG):
     """Run the service."""
-    # Please adapt to package name
-    asyncio.run(run_server(app="my_microservice.__main__:app", config=config))
 
 
 if __name__ == "__main__":
