+++ src/sms/config.py
@@ -13,8 +13,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-This sub-package contains the main business functionality of this service.
+"""Config Parameter Modeling and Parsing."""
 
-It should not contain any service API-related code.
-"""
+from ghga_service_commons.api import ApiConfigBase
+from hexkit.config import config_from_yaml
+from hexkit.log import LoggingConfig
+from pydantic import Field
+
+SERVICE_NAME: str = "sms"
+
+
+@config_from_yaml(prefix=SERVICE_NAME)
+class Config(ApiConfigBase, LoggingConfig):
+    """Config parameters and their defaults."""
+
+    service_name: str = Field(
+        default=SERVICE_NAME, description="Short name of this service"
+    )
+
+
+CONFIG = Config()  # type: ignore
