+++ tests/test_dummy.py
@@ -15,10 +15,7 @@
 
 """Test dummy."""
 
-from my_microservice.core.greeting import generate_greeting
-
 
 def test_dummy():
     """A very simple example test."""
-    greeting = generate_greeting("monde", "French", True)
-    assert greeting.message == "Salut monde!"
+    assert 2 + 2 == 4
